requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
deep-translator==1.11.4
distro==1.9.0
edge-tts==7.0.1
exceptiongroup==1.2.2
filelock==3.18.0
Flask==3.1.0
frozenlist==1.6.0
fsspec==2025.3.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==1.24.4
openai==1.76.2
openai-whisper==20240930
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
srt==3.5.3
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.9.0
torch==2.1.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
yt-dlp==2025.4.30
